Fix Hash Selects return type

+++ Block.h
@@ -28,10 +28,10 @@ public:
   const Record *get(const int idx); // Return ptr to idx record
   uint32_t count();                 // Return number of this->records
 
-  void write(const Record *r, const uint64_t pos=-1);     // Write Record r in this
+  void write(const Record *r, const int pos=-1);     // Write Record r in this
   int read(const uint64_t offset); // Populate the block with N records starting from offset
   void nullify(int reg, int pos, const char* path); // Replaces register in registers[reg] with a bunch of 000's. Then writes to file in pos +- offset
-  void persist(const uint64_t pos=-1); // Write this in disk file and reset this
+  void persist(const int pos=-1); // Write this in disk file and reset this
 };
 
 #endif
